move api types and create skipjob api
