+++ api/v1alpha1/podtypes/access_policy.go
@@ -1,0 +1,50 @@
+package podtypes
+
+// +kubebuilder:object:generate=true
+type AccessPolicy struct {
+	//+kubebuilder:validation:Optional
+	Inbound InboundPolicy `json:"inbound,omitempty"`
+	//+kubebuilder:validation:Optional
+	Outbound OutboundPolicy `json:"outbound,omitempty"`
+}
+
+// +kubebuilder:object:generate=true
+type InboundPolicy struct {
+	//+kubebuilder:validation:Optional
+	Rules []InternalRule `json:"rules"`
+}
+
+// +kubebuilder:object:generate=true
+type OutboundPolicy struct {
+	//+kubebuilder:validation:Optional
+	Rules []InternalRule `json:"rules,omitempty"`
+	//+kubebuilder:validation:Optional
+	External []ExternalRule `json:"external,omitempty"`
+}
+
+type InternalRule struct {
+	//+kubebuilder:validation:Optional
+	Namespace string `json:"namespace,omitempty"`
+	//+kubebuilder:validation:Required
+	Application string `json:"application"`
+}
+
+// +kubebuilder:object:generate=true
+type ExternalRule struct {
+	//+kubebuilder:validation:Required
+	Host string `json:"host"`
+	//+kubebuilder:validation:Optional
+	Ip string `json:"ip,omitempty"`
+	//+kubebuilder:validation:Optional
+	Ports []ExternalPort `json:"ports,omitempty"`
+}
+
+type ExternalPort struct {
+	//+kubebuilder:validation:Required
+	Name string `json:"name"`
+	//+kubebuilder:validation:Required
+	Port int `json:"port"`
+	//+kubebuilder:validation:Required
+	// +kubebuilder:validation:Enum=HTTP;HTTPS;TCP
+	Protocol string `json:"protocol"`
+}
