+++ api/v1alpha1/podtypes/files_env_from.go
@@ -1,0 +1,22 @@
+package podtypes
+
+type EnvFrom struct {
+	//+kubebuilder:validation:Optional
+	ConfigMap string `json:"configMap,omitempty"`
+	//+kubebuilder:validation:Optional
+	Secret string `json:"secret,omitempty"`
+}
+
+type FilesFrom struct {
+	//+kubebuilder:validation:Required
+	MountPath string `json:"mountPath"`
+
+	//+kubebuilder:validation:Optional
+	ConfigMap string `json:"configMap,omitempty"`
+	//+kubebuilder:validation:Optional
+	Secret string `json:"secret,omitempty"`
+	//+kubebuilder:validation:Optional
+	EmptyDir string `json:"emptyDir,omitempty"`
+	//+kubebuilder:validation:Optional
+	PersistentVolumeClaim string `json:"persistentVolumeClaim,omitempty"`
+}
