+++ api/v1alpha1/podtypes/gcp.go
@@ -1,0 +1,11 @@
+package podtypes
+
+type GCP struct {
+	//+kubebuilder:validation:Required
+	Auth Auth `json:"auth"`
+}
+
+type Auth struct {
+	//+kubebuilder:validation:Required
+	ServiceAccount string `json:"serviceAccount"`
+}
