+++ config/rbac/role.yaml
@@ -180,3 +180,13 @@ rules:
   - list
   - update
   - watch
+- apiGroups:
+  - skiperator.kartverket.no
+  resources:
+  - skipjobs
+  - skipjobs/status
+  verbs:
+  - get
+  - list
+  - update
+  - watch
